Generalised to work with any sklearn classfier
(hopefully). Only tested with random forrest so far but they should all have the same interface.

Note that the performance is slightly lower now that we are running the forest with the default arguments.
